query_lambda/requirements.txt
boto3
psycopg2-binary
pgvector
python-dotenv
python-jose
passlib
requests